Rough working version of recursive walk.
Unlike the upstream version, we DON'T link dot directories.  We make new
ones.  This lets you have files that you don't keep track of in the dot
directory (like mutt cache files stored in .mutt).  So the script looks
for the existence of .mutt and creates it if it's not there, then links
the contents of .dotfiles/mutt/ into .mutt/.

Also, I removed the leading underscore thing from all the files names.
I found this unnecessary.